requirements.txt
selenium
behave
allure-behave
allure-pytest
allure-python-commons
pytest-html
faker
pandas
